set durability of weapons and consumables to subtract on use, and consumables to always have durability of 1

+++ core/src/com/counterpart/game/items/Item.java
@@ -8,7 +8,7 @@ public interface Item {
 
     void use();
 
-    double getDurability();
+    int getDurability();
 
 }
